convert commerical module to TS, using $$
